backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
chromadb
openai
tiktoken
pymupdf
langchain-text-splitters
sentence-transformers
torch
